Functions/far/requirements.txt
Flask==2.1.1
requests==2.27.1
asyncio==3.4.3
aiohttp==3.8.1
aioflask==0.4.0
pillow==9.1.0